Dictionaries work as far as having pair vectors of ordinal numbers. Tested on diddy so far.

+++ MasterThesis/RePair_v2/stdafx.h
@@ -45,4 +45,4 @@
 #include "Decoder.h"
 
 //Constants
-const int initialSymbolValue = 256;
+const unsigned long initialSymbolValue = 256;
